Extract CityGML projection model adapter
