+++ src/PlateauResoniteLink/Application/Importing/LocalCityGmlParsedObjectModels.cs
@@ -16,13 +16,6 @@ internal sealed record ParsedRing(
             UVs);
     }
 
-    internal static ParsedRing FromProjectionModel(LocalCityGmlObjectProjection.ParsedRing ring)
-    {
-        return new ParsedRing(
-            ring.RingId,
-            ring.Vertices.Select(GeodeticPoint.FromProjectionModel).ToArray(),
-            ring.UVs);
-    }
 }
 
 internal enum ParsedSurfaceSemantic
@@ -62,18 +55,6 @@ internal sealed record ParsedSurface(
             OpticalProperties);
     }
 
-    internal static ParsedSurface FromProjectionModel(LocalCityGmlObjectProjection.ParsedSurface surface)
-    {
-        return new ParsedSurface(
-            surface.PolygonId,
-            (ParsedSurfaceSemantic)surface.Semantic,
-            ParsedRing.FromProjectionModel(surface.ExteriorRing),
-            surface.InteriorRings.Select(ParsedRing.FromProjectionModel).ToArray(),
-            new ColorRgba(surface.BaseColor.R, surface.BaseColor.G, surface.BaseColor.B, surface.BaseColor.A),
-            surface.TexturePayload,
-            surface.UsesGeneratedDemTexture,
-            surface.OpticalProperties);
-    }
 }
 
 internal sealed record ParsedCityObject(
@@ -113,23 +94,4 @@ internal sealed record ParsedCityObject(
             GeometryHeightMeters);
     }
 
-    internal static ParsedCityObject FromProjectionModel(LocalCityGmlObjectProjection.ParsedCityObject cityObject)
-    {
-        return new ParsedCityObject(
-            cityObject.SlotKey,
-            cityObject.DisplayName,
-            cityObject.PackageName,
-            cityObject.ActualMeshCode,
-            cityObject.LodLevel,
-            cityObject.Surfaces.Select(ParsedSurface.FromProjectionModel).ToArray(),
-            CoordinateReferenceSystem.FromProjectionModel(cityObject.ReferenceSystem),
-            cityObject.SourceFileRelativePath,
-            cityObject.SharedAcrossMeshCodes,
-            cityObject.TerrainAligned,
-            cityObject.GeodeticOriginOverride is null ? null : GeodeticPoint.FromProjectionModel(cityObject.GeodeticOriginOverride),
-            cityObject.FloorsAboveGround,
-            cityObject.MeasuredHeightMeters,
-            cityObject.BuildingAttributes,
-            cityObject.GeometryHeightMeters);
-    }
 }
